requirements.txt
aiogram~=3.7.0
environs~=11.0.0
fastapi~=0.111.0
Werkzeug~=3.0.3
SQLAlchemy~=2.0.30
pydantic~=2.7.1
python-dotenv~=1.0.1
Django~=5.0.3
dnspython~=2.6.1
idna~=3.7
httpcore~=1.0.5
httpx~=0.27.0
anyio~=4.3.0
h11~=0.14.0
PyJWT~=1.7.1
typing_extensions~=4.11.0
pillow~=10.3.0
pip~=24.0
attrs~=23.2.0
contrib~=0.3.0
Jinja2~=3.1.4
rich~=13.7.1
Pygments~=2.18.0
PyYAML~=6.0.1
multidict~=6.0.5
sniffio~=1.3.1
pytz~=2024.1
setuptools~=70.0.0
click~=8.1.7
certifi~=2024.6.2
mdurl~=0.1.2
typer~=0.12.3
shellingham~=1.5.4
asgiref~=3.8.1
sqlparse~=0.4.4
psycopg2~=2.9.9
colorama~=0.4.6
numpy~=2.0.0
MarkupSafe~=2.1.5
orjson~=3.10.5
tablib~=3.5.0
aiohttp~=3.9.5
yarl~=1.9.4
aiosignal~=1.3.1
frozenlist~=1.4.1
python-multipart~=0.0.9
asyncpg~=0.29.0
starlette~=0.37.2
email_validator~=2.1.2
websockets~=12.0
ujson~=5.10.0
uvicorn~=0.29.0
httptools~=0.6.1
watchfiles~=0.22.0
marshmallow~=3.21.3
greenlet~=3.0.3
itsdangerous~=2.2.0
packaging~=24.0
pydantic_core~=2.18.2
python-dateutil~=2.9.0.post0